fix: preserves spaces contained in user config values
