+++ src/lib/utils.js
@@ -40,38 +40,3 @@ export const getPluginSettingsConfig = client => {
     return []
   }
 }
-
-export const generateFlags = (userConfig, nodeSettings) => {
-  if (!Array.isArray(nodeSettings))
-    throw new Error('Settings must be an Array instance')
-  const userConfigEntries = Object.keys(userConfig)
-  let flags = []
-
-  userConfigEntries.forEach(entry => {
-    let flag
-    let configEntry = nodeSettings.find(setting => setting.id === entry)
-    let flagString = configEntry.flag
-
-    if (flagString) {
-      flag = flagString.replace(/%s/, userConfig[entry]).split(' ')
-    } else if (configEntry.options) {
-      const options = configEntry.options
-      const selectedOption = options.find(
-        option =>
-          userConfig[entry] === option.value || userConfig[entry] === option
-      )
-      if (typeof selectedOption['flag'] !== 'string') {
-        throw new Error(
-          `Option "${selectedOption.value ||
-            selectedOption}" must have the "flag" key`
-        )
-      }
-      flag = selectedOption.flag.replace(/%s/, userConfig[entry]).split(' ')
-    } else {
-      throw new Error(`Config entry "${entry}" must have the "flag" key`)
-    }
-    flags = flags.concat(flag)
-  })
-
-  return flags.filter(flag => flag.length > 0)
-}
